+++ src/chart/index.ts
@@ -1,0 +1,2 @@
+export { chart } from "./chart";
+export type { ChartInterface } from "./chart";
