+++ src/chart/y-axis-interaction.ts
